requirements.txt
pydantic>=2.0
pydantic-ai>=0.0.9
openai>=1.0.0
anthropic>=0.18.0
python-dotenv>=1.0.0